Add lock around ClusterWriter.AddIPToSet
The last time we implemented proxy churn metrics, there was a metrics
lock around the call to RecordIPAddress. Since the broker metrics
rewrite, we no longer have this lock. Adding a lock around the call to
AddIPToSet should produce equivalent behaviour. It's not the most
efficient use of locks here but I think it will be sufficient for our
purposes and I've put the call to RecordIPAddress in a goroutine to make
it nonblocking for other broker behaviour.