Generate HMAC key rather than pass it in as arg
Generating the HMAC key for the proxy churn metrics in memory is more
secure than passing it in as a commandline argument. Not only does it
prevent us from accidentally using the default value as in #40161, it
also ensures that the key will not be saved on disk in the systemd file
on the broker machine.

The disadvantage is that our measurements will be partitioned by broker
restarts. We won't be able to compare proxy churn results from before
and after a broker restart.