Rewrite ClusterWriter to store a map of names to sinks
This allows us to define a single ClusterWriter that will write to all
ip count files at once, with the same HMAC key. This allows us to
take the union of counts across pools so we avoid over-counting if the
same IP address appears in different pools within the measurement
window.